pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hexmap"
version = "0.1.0"
description = "Hex map generator"
authors = [
    { name="Mikhail Kartashov", email="kartashov.mik@gmail.com" }
]
dependencies = [
    "pillow",
    "opensimplex"
]
readme = "README.md"
requires-python = ">=3.10"

[project.scripts]
hexmap = "cli:main"  # assumes cli.py has a main() function

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]